+++ Root.tscn
@@ -1,11 +1,8 @@
-[gd_scene load_steps=7 format=3 uid="uid://y7128kiewxq1"]
+[gd_scene load_steps=4 format=3 uid="uid://y7128kiewxq1"]
 
 [ext_resource type="PackedScene" uid="uid://m37vxohbdie7" path="res://SCENES/IconCollection.tscn" id="1_c53fi"]
 [ext_resource type="Script" path="res://CODE/RootWindow.cs" id="1_ymshw"]
-[ext_resource type="PackedScene" uid="uid://bun4786lcjrqy" path="res://SCENES/UI/ArtTitle.tscn" id="2_15dh6"]
-[ext_resource type="PackedScene" uid="uid://c34d7twrflrjj" path="res://SCENES/UI/ArtID.tscn" id="3_gg4vn"]
-[ext_resource type="PackedScene" uid="uid://cff7akcio5o7" path="res://SCENES/StarRating.tscn" id="4_3i65a"]
-[ext_resource type="PackedScene" uid="uid://d2w26aw448b0e" path="res://SCENES/3DView.tscn" id="7_0eqy0"]
+[ext_resource type="PackedScene" uid="uid://mo02sw5hpvph" path="res://SCENES/RightPanel.tscn" id="3_lpm3e"]
 
 [node name="RootWindow" type="Node2D"]
 script = ExtResource("1_ymshw")
@@ -32,25 +29,4 @@ color = Color(0.663665, 0.306272, 0.51105, 0.533333)
 position = Vector2(-553, -301)
 scale = Vector2(2, 2)
 
-[node name="ArtTitle" parent="." instance=ExtResource("2_15dh6")]
-position = Vector2(-86, -304)
-
-[node name="ArtId" parent="." instance=ExtResource("3_gg4vn")]
-position = Vector2(172, -283)
-
-[node name="StarRating" parent="." instance=ExtResource("4_3i65a")]
-position = Vector2(368, -286)
-
-[node name="SubViewportContainer" type="SubViewportContainer" parent="."]
-offset_left = -80.0
-offset_top = -170.0
-offset_right = 1840.0
-offset_bottom = 910.0
-scale = Vector2(0.3, 0.3)
-
-[node name="SubViewport" type="SubViewport" parent="SubViewportContainer"]
-handle_input_locally = false
-size = Vector2i(1920, 1080)
-render_target_update_mode = 4
-
-[node name="3dView" parent="SubViewportContainer/SubViewport" instance=ExtResource("7_0eqy0")]
+[node name="RightPanel" parent="." instance=ExtResource("3_lpm3e")]
