+++ SCENES/RightPanel.tscn
@@ -1,0 +1,70 @@
+[gd_scene load_steps=8 format=3 uid="uid://mo02sw5hpvph"]
+
+[ext_resource type="Script" path="res://SCENES/RightPanel.cs" id="1_6iyby"]
+[ext_resource type="PackedScene" uid="uid://bun4786lcjrqy" path="res://SCENES/UI/ArtTitle.tscn" id="1_qjsnb"]
+[ext_resource type="PackedScene" uid="uid://c34d7twrflrjj" path="res://SCENES/UI/ArtID.tscn" id="2_o2h0e"]
+[ext_resource type="PackedScene" uid="uid://cff7akcio5o7" path="res://SCENES/StarRating.tscn" id="3_p7cer"]
+[ext_resource type="PackedScene" uid="uid://d2w26aw448b0e" path="res://SCENES/3DView.tscn" id="4_uvx0u"]
+
+[sub_resource type="Animation" id="Animation_6f1xb"]
+resource_name = "Grow"
+tracks/0/type = "value"
+tracks/0/imported = false
+tracks/0/enabled = true
+tracks/0/path = NodePath(".:scale")
+tracks/0/interp = 1
+tracks/0/loop_wrap = true
+tracks/0/keys = {
+"times": PackedFloat32Array(0, 1),
+"transitions": PackedFloat32Array(1, 1),
+"update": 0,
+"values": [Vector2(0.3, 0.3), Vector2(0.65, 0.65)]
+}
+tracks/1/type = "value"
+tracks/1/imported = false
+tracks/1/enabled = true
+tracks/1/path = NodePath(".:position")
+tracks/1/interp = 1
+tracks/1/loop_wrap = true
+tracks/1/keys = {
+"times": PackedFloat32Array(0, 1),
+"transitions": PackedFloat32Array(1, 1),
+"update": 0,
+"values": [Vector2(-80, -200), Vector2(-630, -350)]
+}
+
+[sub_resource type="AnimationLibrary" id="AnimationLibrary_w7lco"]
+_data = {
+"Grow": SubResource("Animation_6f1xb")
+}
+
+[node name="RightPanel" type="Node2D"]
+script = ExtResource("1_6iyby")
+
+[node name="ArtTitle" parent="." instance=ExtResource("1_qjsnb")]
+position = Vector2(-86, -304)
+
+[node name="ArtId" parent="." instance=ExtResource("2_o2h0e")]
+position = Vector2(172, -283)
+
+[node name="StarRating" parent="." instance=ExtResource("3_p7cer")]
+position = Vector2(368, -286)
+
+[node name="SubViewportContainer" type="SubViewportContainer" parent="."]
+offset_left = -480.0
+offset_top = -320.0
+offset_right = 1440.0
+offset_bottom = 760.0
+scale = Vector2(0.65, 0.65)
+
+[node name="SubViewport" type="SubViewport" parent="SubViewportContainer"]
+handle_input_locally = false
+size = Vector2i(1920, 1080)
+render_target_update_mode = 4
+
+[node name="3dView" parent="SubViewportContainer/SubViewport" instance=ExtResource("4_uvx0u")]
+
+[node name="AnimationPlayer" type="AnimationPlayer" parent="SubViewportContainer"]
+libraries = {
+"": SubResource("AnimationLibrary_w7lco")
+}
